% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebrel-package.R
\name{create_ebrel_class_object_R}
\alias{create_ebrel_class_object_R}
\title{Create an ebrel data object}
\usage{
create_ebrel_class_object_R(
  E,
  C,
  SD,
  D,
  SxH,
  O,
  LM,
  dim_x,
  dim_y,
  n_h,
  n_s,
  sentinel = 1e+10,
  sigma
)
}
\description{
Create an ebrel data object
}
