% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebrel-package.R
\docType{package}
\name{ebrel-package}
\alias{ebrel}
\alias{ebrel-package}
\title{ebrel: Ebrel - Spatial prioritisation for nature recovery}
\description{
More about what it does (maybe more than one line). Continuation lines should be indented.
}
\author{
\strong{Maintainer}: David Baker \email{d.baker2@exeter.ac.uk} (\href{https://orcid.org/0000-0002-0466-8222}{ORCID})

}
\keyword{internal}
